adding homepage, register, and refactor
